更新代码

+++ scripts/install.js
@@ -5,4 +5,4 @@ const { dirname } = import.meta
 
 const srcPath = resolve(dirname, '../.env')
 const destPath = resolve(dirname, '../.env.development')
-!existsSync(destPath) && copyFileSync(srcPath, destPath)
+existsSync(srcPath) && !existsSync(destPath) && copyFileSync(srcPath, destPath)
